BackEnd/Config/DB/conexiondb.py
#CONEXION DATABASE
from sqlalchemy import create_engine
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker
from dotenv import load_dotenv
import os
#Cargar Datos del .env
load_dotenv()
#Obtener Datos 
DB_HOST = os.getenv("DB_HOST")
DB_PORT = os.getenv("DB_PORT")
DB_NAME = os.getenv("DB_NAME")
DB_USER = os.getenv("DB_USER")
DB_PASSWORD = os.getenv("DB_PASSWORD")
#Crear Ruta
RUTADATABASE = f"mysql+mysqlconnector://{DB_USER}:{DB_PASSWORD}@{DB_HOST}:{DB_PORT}/{DB_NAME}"

engine = create_engine(RUTADATABASE)

try:
    with engine.connect() as connection:
        print("Conexion Exitosa")
except Exception as e:
    print(f"Error al conectar: {e}")

SessionLocal = sessionmaker(autocommit=False,autoflush=False,bind=engine)

Base = declarative_base()